+++ lib/src/types/int.ts
@@ -1,0 +1,2 @@
+/** An integral number. */
+export type Int = number