+++ lib/src/types/integer.ts
@@ -1,2 +1,0 @@
-/** An integral number. */
-export type Integer = number