+++ docs/changelog.md
@@ -4,6 +4,7 @@ Versions and bullets are arranged chronologically from latest to oldest.
 
 ## v4.0.0 (unreleased)
 
+- \[int\]\[millis\] Rename Integer to Int and Milliseconds to Millis
 - \[build\] Use shish-kebab-case for filenames
 - \[build\] Enable NPM workspaces and composite projects
 - \[build\] Upgrade all dependencies